Added a mechanism for trying out different magic hash generation strategies.
This obviously takes a bit of performance hit, but I have come to realize trying out random numbers is not a strategy that takes you very far.
It's good enough to fit n move combinations in tables of size n but for compressing table sizes it's not good enough.
Hopefully, I'll be able to make some progress on that front. It's a difficult nut to crack though.