Interesting and possibly faster way to find pinned pieces using magic hash.
Okay so this is something that I came up with, but I am not sure if anyone has
ever tried. Basically I am using a separate set of magic hashed longs to store
pieces and their pinners. So far as I can think this should work. I still need
to test it, so I can think through on ways to potentially break it.

The idea only works if we already know that we are not under check. Basically,
we use the king's place and a regular rook or bishop mask to do the magic hash
lookup. This returns us a bitboard with the pieces sitting closest to the king
in any direction and the piece that next comes after it. So in every direction
we have exactly two marked set bits. If one of them is an opposite rook/bishop
we can be certain it's behind the other piece (because we're not in check).

Now all we need to check is if the other piece belongs to us. In case it's our
piece we know that it's a pinned piece. Now we can determine if it's a kind of
piece that is partially pinned or a kind that gets fully pinned. Rest of it is
just details. I do hope that I haven't overlooked anything. I am checking this
in because I didn't want to risk it. This remains to be tested. Please leave a
comment if this is not an original idea, or if you think this is one. I'd love
to know all the same.