Documentation fix. Converted link to href.
Basically, I have been trying out a lot of move generation optimizations/tweaks.
I want to match stockfish with Spot/Java, but it has been hard getting past
the current state of efficiency.

Some perft positions this engine seems to on par but on others it's about 25%
slower which is a shame. (11.3 vs 14.1 for instance on one run).

The problem seems to be runtime. I think we can match the speed of Stockfish
given more time, or when allowed to ponder. That is to be verified. But for now
I'll just spend a couple of more months hitting my head against this wall. Or
maybe not. Who knows? I certainly don't.